Issue: UNST-1685 Upgrade prof1D from real to double precision to avoid single precision results for 1D sections in FM.

+++ src/utils_lgpl/gridgeom/packages/gridgeom/src/generalmodules.f90
@@ -250,7 +250,7 @@
    double precision, allocatable     :: dxi   (:)      !< inverse dx
    double precision, allocatable, target :: wu(:)      !< [m] link initial width (m), if < 0 pointer to convtab {"location": "edge", "shape": ["lnx"]}
    double precision, allocatable     :: wui   (:)      !< inverse link initial width (m), if < 0 pointer to convtab
-   real,             allocatable     :: prof1D (:,:)   !< dim = (3,lnx1D) 1= 1D prof width, 2=1D profile height, 3=proftyp, or: if 1,2< 0, pointers to prof 1,2, then 3=alfa1
+   double precision, allocatable     :: prof1D (:,:)   !< dim = (3,lnx1D) 1= 1D prof width, 2=1D profile height, 3=proftyp, or: if 1,2< 0, pointers to prof 1,2, then 3=alfa1
    integer,          allocatable     :: jaduiktmp(:)  !< temparr
    double precision, allocatable, target     :: bob   (:,:)    !< [m] left and right inside lowerside tube (binnenkant onderkant buis) HEIGHT values (m) (positive upward) {"location": "edge", "shape": [2, "lnx"]}
    integer,          allocatable     :: ibot  (:)      !< local ibedlevtype for setting min or max network depths (temporary, result goes to bobs)
